backend/ml_requirements.txt
aiofiles==0.7.0
aiopg==0.15.0
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
clip @ git+https://github.com/openai/CLIP.git@cfcffb90e69f37bf2ff1e988237a0fbe41f33c04
currentsapi==0.0.2
filelock==3.0.12
ftfy==6.0.3
grpcio==1.37.1
grpcio-tools==1.37.1
httptools==0.2.0
huggingface-hub==0.0.8
idna==2.8
joblib==1.0.1
multidict==5.1.0
numpy==1.20.3
packaging==20.9
peewee==3.14.4
peewee-async==0.7.1
Pillow==8.2.0
protobuf==3.17.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
pymilvus==1.1.2
pyparsing==2.4.7
python-dateutil==2.8.0
python-dotenv==0.17.1
regex==2021.4.4
requests==2.21.0
sacremoses==0.0.45
sanic==21.3.4
sanic-routing==0.6.2
six==1.16.0
tokenizers==0.10.3
torch==1.7.1
torchvision==0.8.2
tqdm==4.61.1
transformers==4.6.1
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.24.3
uvloop==0.15.2
wcwidth==0.2.5
websockets==8.1
